error en el r4 :(

+++ it1_A-03_s.canales_s.canales10/src/main/java/uniandes/isis2304/parranderos/negocio/TipoHabitacion.java
@@ -4,11 +4,11 @@ public class TipoHabitacion {
 	
 	private String nombre;
 	
-	private int capacidad;
+	private Integer capacidad;
 	
-	private double costoPorNoche;
+	private Integer costoPorNoche;
 	
-	private int cantidadDisponible;
+	private Integer cantidadDisponible;
 
 	public String getNombre() {
 		return nombre;
@@ -18,7 +18,7 @@ public class TipoHabitacion {
 		this.nombre = nombre;
 	}
 
-	public int getCapacidad() {
+	public Integer getCapacidad() {
 		return capacidad;
 	}
 
@@ -26,23 +26,23 @@ public class TipoHabitacion {
 		this.capacidad = capacidad;
 	}
 
-	public double getCostoPorNoche() {
+	public Integer getCostoPorNoche() {
 		return costoPorNoche;
 	}
 
-	public void setCostoPorNoche(double costoPorNoche) {
+	public void setCostoPorNoche(Integer costoPorNoche) {
 		this.costoPorNoche = costoPorNoche;
 	}
 
-	public int getCantidadDisponible() {
+	public Integer getCantidadDisponible() {
 		return cantidadDisponible;
 	}
 
-	public void setCantidadDisponible(int cantidadDisponible) {
+	public void setCantidadDisponible(Integer cantidadDisponible) {
 		this.cantidadDisponible = cantidadDisponible;
 	}
 
-	public TipoHabitacion(String nombre, int capacidad, double costoPorNoche, int cantidadDisponible) {
+	public TipoHabitacion(String nombre, Integer capacidad, Integer costoPorNoche, Integer cantidadDisponible) {
 		this.nombre = nombre;
 		this.capacidad = capacidad;
 		this.costoPorNoche = costoPorNoche;
